Fix burst mode
Burst mode is only really available with banking. So we switch to banking mode
and need to redefine all port constants.  This eliminates the need for the
second write command, which really previously went to port A.  Note that
we no longer can use the fastest update mode to set custom characters.
With my display, this works up to TWBR=0x05 (615384 Hz).

Speed:  191ms,  38.1x  (full speed)
        144ms,  50.6x  (TWBR = 5)

# Conflicts:
#	Adafruit_RGBLCDShield_Fast.cpp